pyproject.toml
[project]
name = "financeiro"
version = "0.1.0"
description = "Aplicação Django para gestão de contas a pagar e receber."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dj-database-url>=3.0.1",
    "django>=5.2.7",
    "psycopg-binary>=3.2.11",
    "psycopg2-binary>=2.9.11",
    "python-telegram-bot>=22.5",
    "python-dotenv>=1.1.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py311"
select = ["E", "F", "I", "UP"]
ignore = ["E203", "W503"]

[tool.hatch.build.targets.wheel]
packages = [
    "src/financeiro_project",
    "src/entries",
]
